avro,fst,kryo,msgpack adapt to 3.2
