+++ dubbo-serialization-extensions/dubbo-serialization-avro/src/main/java/org/apache/dubbo/common/serialize/avro/AvroObjectInput.java
@@ -32,7 +32,7 @@ import java.util.Map;
 
 public class AvroObjectInput implements ObjectInput {
     private static DecoderFactory decoderFactory = DecoderFactory.get();
-    private BinaryDecoder decoder;
+    private final BinaryDecoder decoder;
 
     public AvroObjectInput(InputStream in) {
         decoder = decoderFactory.binaryDecoder(in, null);
