+++ dubbo-serialization-extensions/dubbo-serialization-common/pom.xml
@@ -38,11 +38,13 @@ limitations under the License.
             <groupId>org.apache.dubbo</groupId>
             <artifactId>dubbo-serialization-api</artifactId>
             <version>3.2.7</version>
+            <optional>true</optional>
         </dependency>
         <dependency>
             <groupId>org.apache.dubbo</groupId>
             <artifactId>dubbo-common</artifactId>
             <version>3.2.7</version>
+            <optional>true</optional>
         </dependency>
     </dependencies>
 </project>
