+++ dubbo-serialization-extensions/dubbo-serialization-fst/src/main/java/org/apache/dubbo/common/serialize/fst/FstObjectInput.java
@@ -29,7 +29,7 @@ import java.lang.reflect.Type;
  */
 public class FstObjectInput implements ObjectInput {
 
-    private FSTObjectInput input;
+    private final FSTObjectInput input;
 
     public FstObjectInput(InputStream inputStream) {
         input = FstFactory.getDefaultFactory().getObjectInput(inputStream);
