+++ dubbo-serialization-extensions/dubbo-serialization-gson/pom.xml
@@ -35,17 +35,14 @@ Licensed to the Apache Software Foundation (ASF) under one or more
         <dependency>
             <groupId>org.apache.dubbo</groupId>
             <artifactId>dubbo-serialization-api</artifactId>
-            <optional>true</optional>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.dubbo</groupId>
-            <artifactId>dubbo-serialization-api</artifactId>
             <version>3.2.7</version>
+            <optional>true</optional>
         </dependency>
         <dependency>
             <groupId>org.apache.dubbo</groupId>
             <artifactId>dubbo-common</artifactId>
             <version>3.2.7</version>
+            <optional>true</optional>
         </dependency>
         <dependency>
             <groupId>org.apache.dubbo.extensions</groupId>
