+++ dubbo-serialization-extensions/dubbo-serialization-jackson/pom.xml
@@ -49,6 +49,7 @@
         <dependency>
             <groupId>org.apache.dubbo</groupId>
             <artifactId>dubbo-common</artifactId>
+            <optional>true</optional>
             <version>3.2.7</version>
         </dependency>
         <dependency>
