+++ dubbo-serialization-extensions/dubbo-serialization-msgpack/pom.xml
@@ -31,6 +31,7 @@ Licensed to the Apache Software Foundation (ASF) under one or more
     <description>The Msgpack serialization implement for dubbo</description>
     <properties>
         <skip_maven_deploy>false</skip_maven_deploy>
+        <dubbo.version>3.2.7</dubbo.version>
     </properties>
 
     <dependencies>
@@ -38,6 +39,14 @@ Licensed to the Apache Software Foundation (ASF) under one or more
         <dependency>
             <groupId>org.apache.dubbo</groupId>
             <artifactId>dubbo-serialization-api</artifactId>
+            <version>${dubbo.version}</version>
+            <optional>true</optional>
+        </dependency>
+
+        <dependency>
+            <groupId>org.apache.dubbo</groupId>
+            <artifactId>dubbo-common</artifactId>
+            <version>${dubbo.version}</version>
             <optional>true</optional>
         </dependency>
