+++ dubbo-serialization-extensions/dubbo-serialization-msgpack/src/main/java/org/apache/dubbo/common/serialize/msgpack/MsgpackObjectOutput.java
@@ -97,7 +97,7 @@ public class MsgpackObjectOutput implements ObjectOutput {
     }
 
     @Override
-    public void writeThrowable(Object obj) throws IOException {
+    public void writeThrowable(Throwable obj) throws IOException {
         writeObject(obj.getClass());
         writeObject(obj);
     }
