+++ dubbo-serialization-extensions/dubbo-serialization-protostuff/pom.xml
@@ -33,12 +33,20 @@ limitations under the License.
 
     <properties>
         <protostuff.version>1.5.9</protostuff.version>
+        <dubbo.version>3.2.7</dubbo.version>
     </properties>
 
     <dependencies>
         <dependency>
             <groupId>org.apache.dubbo</groupId>
             <artifactId>dubbo-serialization-api</artifactId>
+            <version>${dubbo.version}</version>
+            <optional>true</optional>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.dubbo</groupId>
+            <artifactId>dubbo-common</artifactId>
+            <version>${dubbo.version}</version>
             <optional>true</optional>
         </dependency>
         <dependency>
