+++ dubbo-serialization-extensions/dubbo-serialization-test/pom.xml
@@ -28,7 +28,7 @@
     <artifactId>dubbo-serialization-test</artifactId>
     <packaging>jar</packaging>
     <name>${project.artifactId}</name>
-    <version>${revision}</version>
+    <version>3.2.0-SNAPSHOT</version>
     <description>The test serialization module of dubbo project</description>
     <properties>
         <skip_maven_deploy>true</skip_maven_deploy>
@@ -47,11 +47,6 @@
         </dependency>
         <dependency>
             <groupId>org.apache.dubbo.extensions</groupId>
-            <artifactId>dubbo-serialization-jackson</artifactId>
-            <version>${revision}</version>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.dubbo.extensions</groupId>
             <artifactId>dubbo-serialization-protobuf</artifactId>
             <version>1.0.2-SNAPSHOT</version>
             <exclusions>
@@ -62,21 +57,6 @@
             </exclusions>
         </dependency>
         <dependency>
-            <groupId>org.apache.dubbo.extensions</groupId>
-            <artifactId>dubbo-serialization-kryo</artifactId>
-            <version>1.0.2-SNAPSHOT</version>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.dubbo.extensions</groupId>
-            <artifactId>dubbo-serialization-avro</artifactId>
-            <version>1.0.2-SNAPSHOT</version>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.dubbo.extensions</groupId>
-            <artifactId>dubbo-serialization-fst</artifactId>
-            <version>1.0.2-SNAPSHOT</version>
-        </dependency>
-        <dependency>
             <groupId>org.apache.dubbo</groupId>
             <artifactId>dubbo</artifactId>
             <exclusions>
@@ -85,6 +65,16 @@
                     <groupId>com.google.code.gson</groupId>
                 </exclusion>
             </exclusions>
+        </dependency>
+        <dependency>
+            <groupId>org.junit.jupiter</groupId>
+            <artifactId>junit-jupiter-api</artifactId>
+            <version>${junit_jupiter_version}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.hamcrest</groupId>
+            <artifactId>hamcrest</artifactId>
+            <version>${hamcrest_version}</version>
         </dependency>
     </dependencies>
 </project>
