--- dubbo-serialization-extensions/dubbo-serialization-test/src/test/java/org/apache/dubbo/common/serialize/avro/AvroSerializationTest.java
@@ -1,64 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.apache.dubbo.common.serialize.avro;
-
-import org.apache.dubbo.common.serialize.ObjectInput;
-import org.apache.dubbo.common.serialize.ObjectOutput;
-
-import org.hamcrest.Matchers;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.OutputStream;
-
-import static org.apache.dubbo.common.serialize.Constants.AVRO_SERIALIZATION_ID;
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.is;
-import static org.mockito.Mockito.mock;
-
-public class AvroSerializationTest {
-    private AvroSerialization avroSerialization;
-
-    @BeforeEach
-    public void setUp() {
-        this.avroSerialization = new AvroSerialization();
-    }
-
-    @Test
-    public void testContentType() {
-        assertThat(avroSerialization.getContentType(), is("avro/binary"));
-    }
-
-    @Test
-    public void testContentTypeId() {
-        assertThat(avroSerialization.getContentTypeId(), is(AVRO_SERIALIZATION_ID));
-    }
-
-    @Test
-    public void testObjectOutput() throws IOException {
-        ObjectOutput objectOutput = avroSerialization.serialize(null, mock(OutputStream.class));
-        assertThat(objectOutput, Matchers.<ObjectOutput>instanceOf(AvroObjectOutput.class));
-    }
-
-    @Test
-    public void testObjectInput() throws IOException {
-        ObjectInput objectInput = avroSerialization.deserialize(null, mock(InputStream.class));
-        assertThat(objectInput, Matchers.<ObjectInput>instanceOf(AvroObjectInput.class));
-    }
-}
